generating start of copy function
This is the start of a function to generate a copy constructor from
the fake ast into the protobuf.

It is dirty, not elegant, but starts to create the basis for mangled
types functions.

The idea is that generating specific functions instead of using
reflection we can get higher speed.

Runtime should be able to run without reflect, except for typeswitches
on interfaces as needed.